// types/DjangoModelFieldType.ts
export type DjangoModelFieldType =
  | "CharField"
  | "IntegerField"
  | "FloatField"
  | "BooleanField"
  | "DateTimeField"
  | "DateField"
  | "TimeField"
  | "TextField"
  | "ColorField"
  | "JSONField"
  | "FileField"
  | "ImageField"
  | "DecimalField"
  | "ManyToManyField"
  | "ForeignKey"
  | "OneToOneField"
  | "AutoField"
  | "BigAutoField";